feat: add animated landing page
